third_party/WebKit/LayoutTests/fast/html/imports/import-ignore-document-write-expected.txt

CONSOLE WARNING: line 5: Failed to execute 'write' on 'Document': It isn't possible to write into a document from an asynchronously-loaded external script unless it is explicitly opened.
Document.write should be ignored in imports

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


About to call document.write()...
...after calling document.write()
PASS successfullyParsed is true

TEST COMPLETE
